+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - It's now possible to configure a 'Mute override' color which will show a different color when the controlled device is muted
 - Volume change actions now have an option to unmute the device/process
 - The App Finder dialogs have a filter field
+- An option was added to assign a shortcut to profiles to switch between them
 
 **Fix:**